added access control to the site

+++ app/Http/Controllers/PostController.php
@@ -2,11 +2,20 @@
 
 namespace App\Http\Controllers;
 
-use App\Post;
+use App\Post; 
 use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -77,6 +86,10 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id != $post->user_id)
+            return redirect('/posts')->with('error','Unauthorized Page');
+        $this->middleware('auth',['except'=>['index','show']]);
         return view('post.edit')->with('post',$post);
     }
 
@@ -111,6 +124,9 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id != $post->user_id)
+            return redirect('/posts')->with('error','Unauthorized Page');
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
